refactor: API 구조 리팩터링
