+++ src/components/Dashboard/Dashboard.tsx
@@ -2,7 +2,7 @@ import { PlusCircle, Check, ChevronLeft, ChevronRight } from 'lucide-react'
 import { useState, useEffect } from 'react'
 import { DashboardCard } from '../DashboardCard/DashboardCard'
 import { ProjectImportModal } from '../ProjectImportModal'
-import { useProjectList } from '@/api/services/project/index'
+import { useProjectList } from '@/api'
 import { ProjectFormValues } from '../ProjectImportModal/TitleSelect'
 import type { CardType } from '../DashboardCard/DashboardCard'
 import { Button } from '@/components/ui/button'
