+++ src/components/DetailProject/code-selection/CodeViewer.tsx
@@ -2,7 +2,6 @@
 
 import React, { RefObject, useCallback } from 'react'
 import { Textarea } from '@/components/ui/textarea'
-import { Loader2 } from 'lucide-react'
 
 interface CodeViewerProps {
   code: string
